fix handling of ldap CNs with comma also enhanced UI displaying of user names in roles split Ldap.cs into three partial class files closes #2962
